+++ dataset/test/groupby_test.cpp
@@ -587,7 +587,7 @@ TEST_F(GroupbyBinnedTest, sum_with_event_mask) {
 }
 
 TEST_F(GroupbyBinnedTest, concatenate_data_array) {
-  EXPECT_EQ(groupby(a, Dim("labels")).concatenate(Dim::Y), expected);
+  EXPECT_EQ(groupby(a, Dim("labels")).concat(Dim::Y), expected);
 }
 
 TEST_F(GroupbyBinnedTest, concatenate_by_attr) {
@@ -600,7 +600,7 @@ TEST_F(GroupbyBinnedTest, concatenate_by_attr) {
 
 TEST_F(GroupbyBinnedTest, concatenate_data_array_2d) {
   a = bin(a, {makeVariable<double>(Dims{Dim::X}, Shape{2}, Values{1, 8})});
-  auto grouped = groupby(a, Dim("labels")).concatenate(Dim::Y);
+  auto grouped = groupby(a, Dim("labels")).concat(Dim::Y);
   grouped.coords().erase(Dim::X);
   EXPECT_EQ(grouped.slice({Dim::X, 0}), expected);
   // Dim added by grouping is *outer* dim
@@ -612,7 +612,7 @@ TEST_F(GroupbyBinnedTest, concatenate_data_array_conflicting_2d_coord) {
   a.coords().set(
       Dim::X, makeVariable<double>(Dims{Dim::Y, Dim::X}, Shape{3, 3}, units::m,
                                    Values{1, 3, 8, 1, 3, 9, 1, 3, 10}));
-  auto grouped = groupby(a, Dim("labels")).concatenate(Dim::Y);
+  auto grouped = groupby(a, Dim("labels")).concat(Dim::Y);
   EXPECT_EQ(
       grouped.coords().extract(Dim::X),
       makeVariable<double>(Dims{Dim::X}, Shape{2}, units::m, Values{1, 10}));
@@ -622,7 +622,7 @@ TEST_F(GroupbyBinnedTest, concatenate_data_array_conflicting_2d_coord) {
 TEST_F(GroupbyBinnedTest, concatenate_dataset) {
   const Dataset d{{{"a", a}, {"b", a}}};
   const Dataset expected_d{{{"a", expected}, {"b", expected}}};
-  EXPECT_EQ(groupby(d, Dim("labels")).concatenate(Dim::Y), expected_d);
+  EXPECT_EQ(groupby(d, Dim("labels")).concat(Dim::Y), expected_d);
 }
 
 struct GroupbyBinnedMaskTest : public ::testing::Test {
@@ -641,7 +641,7 @@ struct GroupbyBinnedMaskTest : public ::testing::Test {
 };
 
 TEST_F(GroupbyBinnedMaskTest, concatenate) {
-  EXPECT_EQ(groupby(a, Dim("labels")).concatenate(Dim::Y), expected);
+  EXPECT_EQ(groupby(a, Dim("labels")).concat(Dim::Y), expected);
 }
 
 struct GroupbyLogicalTest : public ::testing::Test {